requirements.txt
Flask==2.3.2
requests
python-dotenv
streamlit
bs4
google-generativeai
